Fix DeprecationWarning for float64 casting

+++ python/fbprophet/forecaster.py
@@ -354,8 +354,8 @@ class Prophet(object):
         else:
             # Place potential changepoints evenly through first
             # changepoint_range proportion of the history
-            hist_size = np.floor(
-                self.history.shape[0] * self.changepoint_range)
+            hist_size = int(np.floor(
+                self.history.shape[0] * self.changepoint_range))
             if self.n_changepoints + 1 > hist_size:
                 self.n_changepoints = hist_size - 1
                 logger.info(
